// apps/frontend/src/lib/api/generated/schemas/educationResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Resume Genius API
 * OpenAPI spec version: 1.0.0
 */
import type { DegreeType } from "./degreeType";
import type { EducationResponseFocusArea } from "./educationResponseFocusArea";
import type { EducationResponseStartDate } from "./educationResponseStartDate";
import type { EducationResponseEndDate } from "./educationResponseEndDate";
import type { EducationResponseGpa } from "./educationResponseGpa";
import type { EducationResponseMaxGpa } from "./educationResponseMaxGpa";

/**
 * Response model for education entry.
 */
export interface EducationResponse {
  id: string;
  user_id: string;
  institution_name: string;
  degree: DegreeType;
  field_of_study: string;
  focus_area: EducationResponseFocusArea;
  start_date: EducationResponseStartDate;
  end_date: EducationResponseEndDate;
  gpa: EducationResponseGpa;
  max_gpa: EducationResponseMaxGpa;
  created_at: string;
  updated_at: string;
}