// apps/frontend/src/lib/api/generated/schemas/educationUpdateRequest.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Resume Genius API
 * OpenAPI spec version: 1.0.0
 */
import type { EducationUpdateRequestInstitutionName } from "./educationUpdateRequestInstitutionName";
import type { EducationUpdateRequestDegree } from "./educationUpdateRequestDegree";
import type { EducationUpdateRequestFieldOfStudy } from "./educationUpdateRequestFieldOfStudy";
import type { EducationUpdateRequestFocusArea } from "./educationUpdateRequestFocusArea";
import type { EducationUpdateRequestStartDate } from "./educationUpdateRequestStartDate";
import type { EducationUpdateRequestEndDate } from "./educationUpdateRequestEndDate";
import type { EducationUpdateRequestGpa } from "./educationUpdateRequestGpa";
import type { EducationUpdateRequestMaxGpa } from "./educationUpdateRequestMaxGpa";

/**
 * Request model for updating education entry.
 */
export interface EducationUpdateRequest {
  institution_name?: EducationUpdateRequestInstitutionName;
  degree?: EducationUpdateRequestDegree;
  field_of_study?: EducationUpdateRequestFieldOfStudy;
  focus_area?: EducationUpdateRequestFocusArea;
  start_date?: EducationUpdateRequestStartDate;
  end_date?: EducationUpdateRequestEndDate;
  gpa?: EducationUpdateRequestGpa;
  max_gpa?: EducationUpdateRequestMaxGpa;
}